# app-delivery-backend/src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API delivery",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/restaurante/": {
      "get": {
        "tags": [
          "Restaurante"
        ],
        "summary": "Retorna todos os restaurantes cadastrados.",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista de restaurantes.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id_restaurante": {
                    "type": "number",
                    "example": 1
                  },
                  "nome": {
                    "type": "string",
                    "example": "Pizza Palace"
                  },
                  "email": {
                    "type": "string",
                    "example": "contato@pizzapalace.com"
                  },
                  "endereco": {
                    "type": "string",
                    "example": "Rua das Pizzas, 123"
                  },
                  "telefone": {
                    "type": "string",
                    "example": "11444555666"
                  },
                  "id_usuario": {
                    "type": "number",
                    "example": 3
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Nenhum restaurante foi encontrado.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Nenhum restaurante foi encontrado."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal server error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/restaurante/catalogo": {
      "get": {
        "tags": [
          "Restaurante"
        ],
        "summary": "Retorna o catálogo de pratos de um restaurante específico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID do restaurante",
            "required": true,
            "type": "integer",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Catálogo de pratos do restaurante.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id_prato": {
                    "type": "number",
                    "example": 1
                  },
                  "id_restaurante": {
                    "type": "number",
                    "example": 1
                  },
                  "nome": {
                    "type": "string",
                    "example": "Pizza Margherita"
                  },
                  "descricao": {
                    "type": "string",
                    "example": "Pizza tradicional com molho de tomate, mussarela e manjericão"
                  },
                  "valor": {
                    "type": "number",
                    "example": 3500
                  },
                  "estoque": {
                    "type": "number",
                    "example": 50
                  },
                  "id_categoria": {
                    "type": "number",
                    "example": 6
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "ID do restaurante não fornecido.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Id eh necessario para requisitar o catalogo do resturante."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Restaurante nao tem nenhum prato cadastrado.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Restaurante nao tem nenhum prato cadastrado."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal server error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/restaurante/prato": {
      "post": {
        "tags": [
          "Restaurante"
        ],
        "summary": "Cadastra um novo prato para um restaurante.",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Dados do prato para cadastro.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "id_restaurante": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "Pizza Calabresa"
                },
                "descricao": {
                  "type": "string",
                  "example": "Pizza com molho de tomate, mussarela e calabresa"
                },
                "valor": {
                  "type": "number",
                  "example": 4000
                },
                "estoque": {
                  "type": "number",
                  "example": 100
                },
                "id_categoria": {
                  "type": "number",
                  "example": 6
                }
              },
              "required": [
                "id_restaurante",
                "nome",
                "descricao",
                "valor",
                "estoque",
                "id_categoria"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Prato cadastrado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "id_prato": {
                  "type": "number",
                  "example": 10
                },
                "id_restaurante": {
                  "type": "number",
                  "example": 1
                },
                "nome": {
                  "type": "string",
                  "example": "Pizza Calabresa"
                },
                "descricao": {
                  "type": "string",
                  "example": "Pizza com molho de tomate, mussarela e calabresa"
                },
                "valor": {
                  "type": "number",
                  "example": 4000
                },
                "estoque": {
                  "type": "number",
                  "example": 100
                },
                "id_categoria": {
                  "type": "number",
                  "example": 6
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Dados incompletos para o cadastro do prato.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Todos os campos obrigatorios (id_restaurante, nome, descricao, valor, estoque, id_categoria) devem ser preenchidos."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor durante o cadastro do prato.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal server error"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}